Mister_Moody/requirements.txt
asyncio
hue-api
pyaudio
pydub
python-dotenv
nltk
schedule
SpeechRecognition
wave